Change account setup autodiscovery feature to theme 2
