+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/view/AutoDiscoveryResultHeaderView.kt
@@ -10,13 +10,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextBodyMedium
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextTitleLarge
-import app.k9mail.core.ui.compose.theme.MainTheme
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
+import app.k9mail.core.ui.compose.theme2.MainTheme
 
 @Suppress("LongMethod")
 @Composable
@@ -70,60 +68,5 @@ private fun selectColor(state: AutoDiscoveryResultHeaderState): Color {
         AutoDiscoveryResultHeaderState.NoSettings -> MainTheme.colors.primary
         AutoDiscoveryResultHeaderState.Trusted -> MainTheme.colors.success
         AutoDiscoveryResultHeaderState.Untrusted -> MainTheme.colors.warning
-    }
-}
-
-@Preview
-@Composable
-internal fun AutoDiscoveryResultHeaderViewTrustedCollapsedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultHeaderView(
-            state = AutoDiscoveryResultHeaderState.Trusted,
-            isExpanded = true,
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun AutoDiscoveryResultHeaderViewTrustedExpandedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultHeaderView(
-            state = AutoDiscoveryResultHeaderState.Trusted,
-            isExpanded = false,
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun AutoDiscoveryResultHeaderViewUntrustedCollapsedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultHeaderView(
-            state = AutoDiscoveryResultHeaderState.Untrusted,
-            isExpanded = true,
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun AutoDiscoveryResultHeaderViewUntrustedExpandedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultHeaderView(
-            state = AutoDiscoveryResultHeaderState.Untrusted,
-            isExpanded = false,
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun AutoDiscoveryResultHeaderNoSettingsPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultHeaderView(
-            state = AutoDiscoveryResultHeaderState.NoSettings,
-            isExpanded = false,
-        )
     }
 }
