+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/view/AutoDiscoveryResultView.kt
@@ -10,18 +10,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import app.k9mail.autodiscovery.api.AuthenticationType
 import app.k9mail.autodiscovery.api.AutoDiscoveryResult
-import app.k9mail.autodiscovery.api.ConnectionSecurity
-import app.k9mail.autodiscovery.api.ImapServerSettings
-import app.k9mail.autodiscovery.api.SmtpServerSettings
-import app.k9mail.core.common.net.toHostname
-import app.k9mail.core.common.net.toPort
 import app.k9mail.core.ui.compose.designsystem.atom.Surface
-import app.k9mail.core.ui.compose.theme.MainTheme
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
+import app.k9mail.core.ui.compose.theme2.MainTheme
 
 @Composable
 internal fun AutoDiscoveryResultView(
@@ -70,61 +62,5 @@ internal fun AutoDiscoveryResultView(
                 }
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun AutoDiscoveryResultViewTrustedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultView(
-            settings = AutoDiscoveryResult.Settings(
-                incomingServerSettings = ImapServerSettings(
-                    hostname = "imap.example.com".toHostname(),
-                    port = 993.toPort(),
-                    connectionSecurity = ConnectionSecurity.TLS,
-                    authenticationTypes = listOf(AuthenticationType.PasswordEncrypted),
-                    username = "",
-                ),
-                outgoingServerSettings = SmtpServerSettings(
-                    hostname = "smtp.example.com".toHostname(),
-                    port = 465.toPort(),
-                    connectionSecurity = ConnectionSecurity.TLS,
-                    authenticationTypes = listOf(AuthenticationType.PasswordEncrypted),
-                    username = "",
-                ),
-                isTrusted = true,
-                source = "preview",
-            ),
-            onEditConfigurationClick = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun AutoDiscoveryResultViewUntrustedPreview() {
-    PreviewWithThemes {
-        AutoDiscoveryResultView(
-            settings = AutoDiscoveryResult.Settings(
-                incomingServerSettings = ImapServerSettings(
-                    hostname = "imap.example.com".toHostname(),
-                    port = 993.toPort(),
-                    connectionSecurity = ConnectionSecurity.TLS,
-                    authenticationTypes = listOf(AuthenticationType.PasswordEncrypted),
-                    username = "",
-                ),
-                outgoingServerSettings = SmtpServerSettings(
-                    hostname = "smtp.example.com".toHostname(),
-                    port = 465.toPort(),
-                    connectionSecurity = ConnectionSecurity.TLS,
-                    authenticationTypes = listOf(AuthenticationType.PasswordEncrypted),
-                    username = "",
-                ),
-                isTrusted = false,
-                source = "preview",
-            ),
-            onEditConfigurationClick = {},
-        )
     }
 }
